Created a basic servlet for the game
