+++ src/edu/ycp/cs320/assign01/servlet/TextBasedServlet.java
@@ -1,0 +1,69 @@
+package edu.ycp.cs320.assign01.servlet;
+
+import java.io.IOException;
+
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+public class TextBasedServlet extends HttpServlet {
+	private static final long serialVersionUID = 1L;
+
+	@Override
+	protected void doGet(HttpServletRequest req, HttpServletResponse resp)
+			throws ServletException, IOException {
+
+		System.out.println("TextBased Servlet: doGet");	
+		
+		// call JSP to generate empty form
+		req.getRequestDispatcher("/_view/textBased.jsp").forward(req, resp);
+	}
+	
+	@Override
+	protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+			throws ServletException, IOException {
+		
+		System.out.println("TextBased Servlet: doPost");
+		
+		String errorMessage = null;
+		String output = null;
+		String input = req.getParameter("input");
+		// TextBasedModel model;
+		// if(req.getAttribute("game") != null)
+		// 	model = req.getAttribute("game");
+		// else
+		//	model = new TextBasedModel();
+		// TextBasedController controller = new TextBasedController(model);
+
+		// check for errors in the form data before using is in a calculation
+		if (input == null || input.equals("")) {
+			errorMessage = "Please specify input";
+		}
+		// otherwise, data is good, do the calculation
+		// must create the controller each time, since it doesn't persist between POSTs
+		// the view does not alter data, only controller methods should be used for that
+		// thus, always call a controller method to operate on the data
+		else {
+			// output = controller.getOutput(input);
+			output = "You entered \"" + input + "\"";
+		}
+		
+		// Add parameters as request attributes
+		// this creates attributes named "first" and "second for the response, and grabs the
+		// values that were originally assigned to the request attributes, also named "first" and "second"
+		// they don't have to be named the same, but in this case, since we are passing them back
+		// and forth, it's a good idea
+		
+		//req.setAttribute("game", model);
+		
+		// add result objects as attributes
+		// this adds the errorMessage text and the result to the response
+		req.setAttribute("errorMessage", errorMessage);
+		// req.setAttribute("result", result);
+		req.setAttribute("output", output);
+		
+		// Forward to view to render the result HTML document
+		req.getRequestDispatcher("/_view/textBased.jsp").forward(req, resp);
+	}
+}
