+++ war/_view/index.jsp
@@ -9,7 +9,7 @@
 		Welcome to the index view jsp! Click one of the following buttons to go to that page.
 		<form action="${pageContext.servletContext.contextPath}/index" method="post">
 			<input type="Submit" name="add" value="Add Numbers">
-			<input type="Submit" name="multiply" value="Multiply Numbers">
+			<input type="Submit" name="text" value="Text Based">
 		</form>
 	</body>
 </html>
